+++ src/App/Documentation/index.js
@@ -1,13 +1,10 @@
 import React from "react";
-import { withRouter, Switch, Route, Redirect } from "react-router-dom";
+import { Switch, Route, Redirect } from "react-router-dom";
 
 import SelectPanel from "./SelectPanel";
 import DocHeading from "./utils/DocHeading";
 import RenderRoutes from "./utils/RenderRoutes";
 import routes from "../routes/docs";
-
-const SelectPanelWithRoutes = withRouter(props => <SelectPanel {...props} />);
-const DocHeadingWithRoutes = withRouter(props => <DocHeading {...props} />);
 
 const Documentation = () => (
     <div className="documentation">
@@ -16,11 +13,11 @@ const Documentation = () => (
                 <h1>DesignGuide</h1>
             </div>
             <div className="view-heading col-md-9 col-sm-12">
-                <DocHeadingWithRoutes />
+                <DocHeading />
             </div>
         </div>
         <div className="row">
-            <SelectPanelWithRoutes />
+            <SelectPanel />
             <main className="doc-view col-md-9 col-sm-12">
                 <Switch>
                     <Route exact path="/docs" render={() => <Redirect to="/docs/getting-started" />} />