+++ src/App/components/Loader/index.js
@@ -2,20 +2,12 @@ import React from "react";
 import PropTypes from "prop-types";
 
 const Loader = ({ id, size, muted, visible }) => {
-    return (
-        <div id={id} className={`loader loader-${size}${muted ? " loader-muted" : ""}${visible ? " in" : ""}`}>{"\n\t"}
-            <ul className="loader-icon">{"\n\t\t"}
-                <li></li>{"\n\t\t"}
-                <li></li>{"\n\t\t"}
-                <li></li>{"\n\t"}
-            </ul>{"\n"}
-        </div>
-    );
+    return <div id={id} className={`loader loader-${size}${muted ? " loader-muted" : ""}${visible ? " in" : ""}`} data-loader="true"></div>;
 };
 
 Loader.propTypes = {
     id: PropTypes.string,
-    size: PropTypes.string.isRequired,
+    size: PropTypes.oneOf(["small", "default", "large"]).isRequired,
     muted: PropTypes.bool,
     visible: PropTypes.bool
 };