+++ src/App/index.js
@@ -1,25 +1,34 @@
 import React from "react";
-import { BrowserRouter, HashRouter, withRouter, Switch, Route } from "react-router-dom";
+import { Router, Switch, Route } from "react-router-dom";
+import createBrowserHistory from "history/createBrowserHistory";
 
 import routes from "./routes/root.js";
 import AppHeader from "./AppHeader";
 import ErrorPage404 from "./ErrorPage404";
 
-const Router = (process.env.NODE_ENV === "production") ? HashRouter : BrowserRouter;
-
-const AppHeaderWithRoutes = withRouter(props => <AppHeader {...props} />);
+// const isProd = process.env.NODE_ENV === "production";
+// const history = createBrowserHistory({ basename: "/design.payex.com" });
+const history = createBrowserHistory();
 
 const App = () => (
-    <Router>
+    <Router history={history}>
         <div id="px-designguide">
-            <AppHeaderWithRoutes />
+            <AppHeader />
             <Switch>
                 {routes.map((route, i) => {
                     const { path, component, exact } = route;
                     const RouteComponent = component.default;
 
                     return (
-                        <Route key={i} exact={exact} path={path} render={() => <RouteComponent />} />
+                        <Route key={i} exact={exact} path={path} render={() => {
+                            history.listen(location => {
+                                window.gtag("config", "UA-3440932-20", {
+                                    "page_location": window.location.href,
+                                    "page_path": location.pathname
+                                });
+                            });
+                            return <RouteComponent />;
+                        }} />
                     );
                 })}
                 <Route component={ErrorPage404} />