+++ src/px-script/datepicker/formats.js
@@ -1,0 +1,71 @@
+// Languages codes are specified in ISO 639-1
+// https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+
+export default {
+    nb: {
+        dateFormat: "DD.MM.YYYY",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["januar", "februar", "mars", "april", "mai", "juni", "juli", "august", "september", "oktober", "november", "desember"],
+            monthsShort: ["jan", "feb", "mar", "apr", "mai", "jun", "jul", "aug", "sep", "okt", "nov", "des"],
+            weekdays: ["søndag", "mandag", "tirsdag", "onsdag", "torsdag", "fredag", "lørdag"],
+            weekdaysShort: ["søn", "man", "tir", "ons", "tor", "fre", "lør"],
+            weekdaysMin: ["sø", "ma", "ti", "on", "to", "fr", "lø"]
+        }
+    },
+    sv: {
+        dateFormat: "YYYY-MM-DD",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["januari", "februari", "mars", "april", "maj", "juni", "juli", "augusti", "september", "oktober", "november", "december"],
+            monthsShort: ["jan", "feb", "mar", "apr", "maj", "jun", "jul", "aug", "sep", "okt", "nov", "dec"],
+            weekdays: ["söndag", "måndag", "tisdag", "onsdag", "torsdag", "fredag", "lördag"],
+            weekdaysShort: ["sön", "mån", "tis", "ons", "tor", "fre", "lör"],
+            weekdaysMin: ["sö", "må", "ti", "on", "to", "fr", "lö"]
+        }
+    },
+    da: {
+        dateFormat: "DD.MM.YYYY",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["januar", "februar", "marts", "april", "maj", "juni", "juli", "august", "september", "oktober", "november", "december"],
+            monthsShort: ["jan", "feb", "mar", "apr", "maj", "jun", "jul", "aug", "sep", "okt", "nov", "dec"],
+            weekdays: ["søndag", "mandag", "tirsdag", "onsdag", "torsdag", "fredag", "lørdag"],
+            weekdaysShort: ["søn", "man", "tir", "ons", "tor", "fre", "lør"],
+            weekdaysMin: ["sø", "ma", "ti", "on", "to", "fr", "lø"]
+        }
+    },
+    fi: {
+        dateFormat: "DD.MM.YYYY",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["tammikuu", "helmikuu", "maaliskuu", "huhtikuu", "toukokuu", "kesäkuu", "heinäkuu", "elokuu", "syyskuu", "lokakuu", "marraskuu", "joulukuu"],
+            monthsShort: ["tammi", "helmi", "maalis", "huhti", "touko", "kesä", "heinä", "elo", "syys", "loka", "marras", "joulu"],
+            weekdays: ["sunnuntai", "maanantai", "tiistai", "keskiviikko", "torstai", "perjantai", "lauantai"],
+            weekdaysShort: ["su", "ma", "ti", "ke", "to", "pe", "la"],
+            weekdaysMin: ["su", "ma", "ti", "ke", "to", "pe", "la"]
+        }
+    },
+    en: {
+        dateFormat: "DD/MM/YYYY",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"],
+            monthsShort: ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
+            weekdays: ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"],
+            weekdaysShort: ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"],
+            weekdaysMin: ["Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"]
+        }
+    },
+    iso8601: {
+        dateFormat: "YYYY-MM-DD",
+        hourFormat: "HH:mm",
+        locale: {
+            months: ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"],
+            monthsShort: ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
+            weekdays: ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"],
+            weekdaysShort: ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"],
+            weekdaysMin: ["Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"]
+        }
+    }
+};